Store data in Postgres
This add support for postgres and stores all the data in a DB. Check out `.env.example` for list of environment variables you can set.

To create a new database and populate it with seeds, call:

```sh
$ scripts/build-database.sh
```

This also adds some *very* lightweight authorization for API requests that change data or want to read private data. In production you'll need to include the `API_KEYS` environment variable to support requests to update data.