Add demo UI (#33)
The UI lives in the `ui` folder. You can start a process build & watch the source with `npm start` or just build with `npm run build`. The main idea here is that all the source is bundled into a single JS file so it's easy to drop directly onto a web page (no need for an iframe, for example). We should probably do some more to isolate CSS selectors, though.