Add an SSL listener for our load balancers (#35)
I'm not sure this will actually work since there's not a certificate specified here. But maybe, since we are using the `amazonaws.com` domain.  We need to enable SSL support on the API in order for the front-end demo to work, otherwise data won’t load because the front-end is HTTPS and the API is HTTP.