terraform: add https (#36)
* terraform: add https

This commit updates the TF setup with a couple of key variables

- domain_name: if supplied will connect the DNS to the LB
- ssl_enabled: if true, will add an HTTPS listener

Here, I've created the DNS records out-of-line, since I think
verifying is a bit easier outside of TF, but we can pull in
the created ids and use them here.

An important note for terraform: the `count` field governs
how many of the resource are created. If the count is 0, nothing is
created. The `data` type pulls info from AWS, instead of managing
that resource directly.