Add multi-state support for VaccineSpotter loader
Use the `--states` (or `--vaccinespotter-states`) option to specify a list of states to load data for in multi-state sources (e.g. cvsApi, vaccinespotter).

In VaccineSpotter, we're a little conservative. Because we want to form IDs that work across sources, we implement a custom formatter for each provider/brand. I've implemented all the onles VaccineSpotter currently supports in Colorado and New Jersey, but other states may miss some VaccineSpotter locations (but we log those as warnings).