Support colorado in vaccinespotter loader (#39)
Use the `--states` (or `--vaccinespotter-states`) option to specify a list of states to load data for in multi-state sources (only supported in `vaccinespotter` in this PR).

This also expands the VaccineSpotter source to pull in more than Walgreens information. Because it covers a variety of sources that may need special handling (e.g. custom IDs, special location types), this adds a mechanism for customizing how the data gets formatted by provider and/or brand.