ui: adding private flag
This commit adds the API key and private flag to pull in data.

Note that with the current UI, this value is exposed to the
frontend. I'm not sure how big of a deal this is, but if we
really wanted it to be 'secret', we might want to re-structure
from a webpack app, or setup some intermediate proxy.