Limit concurrent updates sent from loader
The loader often finds results in batches or hundreds or thousands, but it's not reasonable to send all those in parallel to the database. Instead limit concurrency to just a few at a time. By default, the loader will send 10 updates at a time, but you can adjust that by setting the `API_CONCURRENCY` environment variable. For example, set:

    export API_CONCURRENCY=20

For 20 concurrent updates, or:

    export API_CONCURRENCY=1

To send updates serially.

Fixes #60.