Support the SMART Scheduling Links standard (#61)
Adds basic support for the SMART Scheduling Links standard (https://github.com/smart-on-fhir/smart-scheduling-links). The API provides read access to appointment availability information and can be accessed at `/smart-scheduling/$bulk-publish` (all related endpoints are under `/smart-scheduling/*`).

This is not necessarily a *good* implementation of the standard — as an aggregator, we often lack detailed enough information about availability to really provide the information the standard is designed for, so there are some funny bits around “unknown” availability and capacity at a location:
- Locations each only have one schedule and one slot (we could provide multiple schedules for locations where we have more details, but in this initial implementation we don’t).
- When availability is unknown, the slot has `status = "free"` but `capacity = 0`.
- When availability is known, the slot has `capacity = 1` (rather than an actual count) and the appropriate `status`.
- The dates for the slot are complete nonsense — they’re just 24 hours in the future from whenever you request, since we often don’t know anything about available dates. (Again, we could be more detailed in cases where we know more, but are starting with the common denominator all locations right now, which is no dates at all.)

Ultimately, this is a decent strawman for the SMART Scheduling Connectathon, but not necessarily a great implementation.

This also makes some general fixes to error handling.

Fixes #54.