Limit concurrent updates sent from loader (#64)
The loader often finds results in batches or hundreds or thousands, but it's not reasonable to send all those in parallel to the database. Instead limit concurrency to just a few at a time. By default, the loader will send 10 updates at a time, but you can adjust that by setting the `API_CONCURRENCY` environment variable.

For example, to send 20 updates at a time, set:

    $ export API_CONCURRENCY=20
    $ bin/appointment-availability-loader.js --send cvsApi

Or to send updates serially:

    $ export API_CONCURRENCY=1
    $ bin/appointment-availability-loader.js --send cvsApi

Fixes #60.