Add a server mode to loader, which runs as an HTTP server (#73)
You can now run the loader as a web server. When you POST to `/`, it will run whatever sources you specify. Use the `server` subcommand to start it:

```sh
$ bin/appointment-availability-loader server [additional options]
```

Then send a POST request to it with any of the normal command-line options as a JSON object. For example, to run the VaccineSpotter and NJVSS sources for New Jersey:

```sh
$ curl 'http://localhost:3010/' --data '{ "sources": ["njvss", "vaccinespotter"], "states": "NJ" }'
```

You can set the port the server listens on with the `PORT` environment variable.