Actually parse Rite Aid `last_updated` dates (#80)
Due to a horrible copy/paste error on my part, we were checking the format of the actual data string for Rite Aid, but then *parsing* a hard-coded value! Needless to say, our dates were not very correct. This also improves parsing by trying to parse the date as ISO 8601 if it was not in Rite Aid's wonky format, and if that fails, logging a useful error.