Update locations and availability by `external_ids` (#102)
Make updates work by looking up locations using the sent `id` (just like today), but if no location is found with that ID, it looks up locations by each of the specified `external_ids` instead (not like today!). If `id` is not sent, it’ll just skip to that second step.

This is the way `external_ids` was always intended to function, and doing this also paves the way for #98, or any situation where `id` is not determined by the data source, but is instead a unique identifier in our system.

Fixes #48.