Set proper `external_ids` for NJVSS data (#111)
This is a somewhat significant refresh of our NJVSS loader. It pulls in new updates from the New Jersey team, and fixes #101 and #31.

- Set `external_ids` for NJVSS.
    - All entries have `njiis_covid` where the value is `"<IIS_ID>:<Name>"`.
    - Walmart has some entries in this dataset, and those also have `walmart` IDs with the store number.
- Added vaccine product information (mainly just porting over the NJ team’s work)
- NJVSS sites no longer require waitlist registration.
- Only use the description if it’s meaningful (when no appointments are available, the description in the data is just a message saying there are no appointments, rather than useful info about the location).
- Update the location information in the DB every time we run this, since we need to get these new external IDs in. As a bonus, this fixes #31. (We’ll have to revisit this when we start filling in the location table with more human-curated info from NJ or from Vaccinate the States).

Note: Name is in the `njiis_covid` external identifier because some clinics run multiple locations that are somewhat ad-hoc and do not exist as separate entries in the state IIS. This is somewhat analogous to the issue with VTrckS PINs. We record the *actual* NJIIS ID in `location.meta.njiis`, since it’s not unique (we might change that in the future).

Fixes #31. Fixes #101.