HOTFIX: send correct arguments to loader handler (#112)
In #111, we modified the `handler` callback in the loader to optionally take a second argument. However, we previously used `[].forEach(handler)` in a bunch of places, which happens to send a second argument (of the wrong type!) that was previously ignored, and now causes problems. This adds a clearer error and fixes all the known cases of this issue.

This also fixes a couple server bugs I tripped over figuring this out.