Switch to randomized ids for locations (#113)
This changes the `id` column in the `provider_locations` table to a UUID, and fixes #98. (See that issue for more details on the reasoning here.) The old string ID is preserved for posterity in `id_old` and also surfaced in the API as an entry in `external_ids` named `univaf_v0`.

Fixes #98.