HOTFIX: handle updates with IDs that are not UUIDs
Old loaders are still sending updates with IDs. That's OK, or at least it was supposed to be. Turned out I did not test this scenario after we switched from integers to UUIDs, and now we get Postgres errors when the ID we used could not be converted to an actual UUID. This fixes the issue.