Change query to better use new indexes
Poking around in SQL with EXPLAIN ANALYZE:
- Old query w/ old indexes: (cost=3032.72..3034.93 rows=883 width=659) (actual time=659.867..660.105 rows=883 loops=1)
- Old query w/ new indexes: (cost=815.76..817.97 rows=883 width=659) (actual time=47.751..47.887 rows=883 loops=1)
- New query w/ new indexes: (cost=593.44..594.23 rows=316 width=659) (actual time=23.567..23.694 rows=882 loops=1)

(All from my local database with a little under 1k rows, and `SET enable_seqscan = OFF`)

I originally wrote the old query with the subquery instead of a CTE because it seemed simpler to grasp, and the performance was similar. With some more thinking about the indexes, h
owever, the CTE gets a lot faster.