Optimize locations query and respond in batches (#124)
This is an intermediary step along the way to #65 (pagination). We can’t immediately turn on pagination because it would change the result format, and we do have current users, so this causes our current `/locations` query to stream out results in batches of 2000. As a next PR, paginated URLs will be at the new `/api/edge/locations` URL (see #90) so they are opt-in. This at least makes the current mega-query safer as we pump more data into the system.

Along the way, this makes a HUGE optimization to the locations query, partially by changing the query and partially by adding some useful indexes. I also changed the sorting to be by `created_at` instead of `updated_at` — that way the index for it stays a more stable over time. (Note that this is actually `(created_at, id)` since the timestamp is not unique and so does not guarantee sortability. I forgot about this in #113, and it would have been a better reason to use auto-incrementing IDs, but that ship has sailed for now.)

Performance note son the query… using `EXPLAIN ANALYZE` on my local DB (just under 1k rows, but with `SET enable_seqscan = OFF` to simulate a bigger DB):

- Old query w/ old indexes: (cost=3032.72..3034.93 rows=883 width=659) (actual time=659.867..660.105 rows=883 loops=1) (YOWZA! 😱)
- Old query w/ new indexes: (cost=815.76..817.97 rows=883 width=659) (actual time=47.751..47.887 rows=883 loops=1)
- New query w/ new indexes: (cost=593.44..594.23 rows=316 width=659) (actual time=23.567..23.694 rows=882 loops=1)