HOTFIX: multi-batch streaming of locations breaks
When I cleaned up the way batching was done in #124, I apparently made it possible for the query to have ambiguous references that sometimes fail (not entirely sure what the conditions are for this, since it is clearly working sometimes). This fully qualifies the pagination key references, which should resolve the issue.