Add top-level detail fields to `availability` (#121)
This adds several fields that were previously buried in `meta` to the top level in `availability`:
- `availability_count`: number of open appointments
- `products`: array of vaccine products that are avaialable at the location (e.g. `["moderna", "pfizer"]`)
- `doses`: array of doses that are available at the location (e.g. `["all_doses", "second_dose_only"]`)
- `capacity`: details of available appointments/products/doses by day
- `slots`: as finely detailed as possible info about availabel appointments/products/doses. Ideally, this is every open slot, but could be more coarse-grained by time (e.g. each item in this array might cover a morning or afternoon, or a 1-hour timeframe, or whatever).

Note any or all of the above could be NULL. These are meant to make the information easier to access in a more standard form, when it's available.

The bulk of the work here is in validating the format of `capacity` and `slots`. We store them as arrays of JSONB data in Postgres, so the DB can’t do all the type validation work (and even if it did, not would probably provide less clear messaging about where the malformed item in your data is). I did this because the entire set of slots/capacity entries is unique to a single availability update, and if we had separate tables and relationships for these, every update would have to delete and recreate lots of rows. It seemed useful to make these JSONB so they are more “atomic,” in a sense.