Create loader for CVS SMART Scheduling Links API (#134)
CVS's new SMART Scheduling Links API includes slightly more detailed data than their previous API, and is open access (yay!). We can use it to cover the whole country for CVS.

Note that this is set to log New Jersey results as private, and is still running alongside the old, closed-access CVS API. It turns out we’re getting different results for a lot of locations with this API (vs. the old one), so we’re making it non-production in NJ until we better understand the situation (it’s public for all other states, since we don’t have any data for them otherwise).

Fixes #77.