Simplify the `getLocationByExternalIds()` query (#141)
We only call `getLocationsByExternalIds()` when trying to find the location to update in the `/update` endpoint. Unfortunately, that query calls down into `getLocations()`, which runs a very complicated query with a CTE that does a LOT more work than we need for this case -- and this case needs to be optimized because we get updates happening *constantly*. This simplifies the query to just look at the `provider_locations` table, which should make it much lighter and faster.

There's more we can do here in terms of moving `external_ids` into a separate table, and possibly seeing what else in the `/update` path can be optimized, but this seems like a good start.