Handle VaccineSpotter capacity-style appointments
It turns out VaccineSpotter's `appointments` field can have two differently formatted types of values, and we were only expecting one format. (We expected it to have specific slot times, analogous to how we format individual slots, but it can also have dates, which are analogous to how we handle capacity). We now account for both formats, which should resolve some errors we've been seeing *thousands* of in Sentry. (I've also confirmed the possible formats with @GUI.)