Add script to remove duplicate locations (#157)
This is a quick-n-dirty script to merge duplicate provider locations in the database. Running it will identify duplicates and print out a list of the deletions and changes it would make, but not actually make any changes in the database. Run it with the `--commit` to actually change the database.

It also prints a JSON object describing what location ID should replace ones that were removed if needed for later reconciling.