Restrict what states are supported for Rite Aid (#167)
Restrict what states we check for Rite Aid to only those where there are Rite Aid stores. Their API is giving us occasional 403 responses because we make too many requests. The way I've done this is a little redundant (the list of states is limited in the runtime configuration, but *also* in the source), but the source-level restriction seemed useful in case we are ever less careful than we should be in the future.

We used to query an even smaller set of locations (based on a map of where vaccinations are offered on Rite Aid's COVID-19 page), but later found some results outside that list, so expanded it to *all* states. This tries to cut a better happy medium by querying all states where there are Rite Aid stores, which means fewer queries, but should still cover a superset of everywhere vaccinations might be offered.