Don't log full availability if it didn't change
We previously logged a complete `availability` record every time a new availability update was sent to the DB. That's great for deep auditing, but it's a lot of redundant data when it comes to pretty much any other use case. This led to our database running out of space really quickly, but also makes it harder for consumers of our data to make good use of it because it's so unnecessarily big.

This lightens the load a little by only logging the `checked_at` field if that's the only thing that changed, and only logging `checked_at` and `valid_at` if they changed but none of the other data did.