Warn and do not save 4-digit zip codes from NJVSS
NJVSS has some malformed records with 4-digit zip codes in the address. These are often missing a `0` prefix (so they probably got parsed as a number somewhere), but not always, so we can't automatically fix them. Instead, we warn on the situation and simply do not save the zip code (but do save the rest of the address).