Get product/dose information even when no slots
Like other sources, VaccineSpotter does not always have slot-level info, but still strives to surface product and dose information *overall* even when it's not available by slot or date. We were only loading the info if VaccineSpotter had it at the slot level. This falls back to loading it at the top level if not available by slot.