Handle bad njvss zip codes (#193)
NJVSS has some malformed records with 4-digit zip codes in the address. These are often missing a `0` prefix (so they probably got parsed as a number somewhere), but not always, so we can't automatically fix them. Instead, we warn on the situation and simply do not save the zip code (but do save the rest of the address).

While I was at it, I also added `njvss_res_id` as an external ID.  NJVSS now outputs `res_id`, which I understand is the GUID for the ID of the location in Microsoft Dynamics, which is what Microsoft VRAS is built on, which is what NJVSS is built on.