Add `merge-locations` script
This script can be used to manually merge several specified locations together. It supports situations where the `remove-duplicate-locations` script cannot automatically identify duplicates that should be merged because they do not share external IDs -- for example, an NJVSS location and VaccineSpotter location that are the same, but the NJVSS record only has an NJVSS ID, while the VaccineSpotter one only has other IDs.

List the IDs of locations to merge as arguments, where the first ID is the one others will be merged into:

    $ scripts/merge-locations.js <id_1> <id_2> <id_3> ...etc...

Like the `remove-duplicate-locations` script, this will print a plan, but not actually merge unless you pass `--commit` as an argument.

This also rewrites `remove-duplicate-locations` so that these two scripts can share much of their code, and should hopefully make that script easier to follow.