Remove old logs from the database when dumping (#184)
Start removing old availability log entries from the database when we write them out to S3. The current logging script looks to be working extremely well, so we can start destroying old entries. (This also changes things to delete logs more than 7 days old, so we keep them around for a bit in case we need them to remediate issues.)

Co-authored-by: Aston Motes <astonm@gmail.com>