Keep track of when availability was last changed (#181)
This adds a new field to `availability` called `changed_at`, which tracks the last time the availability data changed (it is not affected by the `checked_at` and `valid_at` timestamps). When we log availability changes, we always include it.

This is a follow-on from https://github.com/usdigitalresponse/appointment-availability-infra/pull/176#issuecomment-854906497. Fixes #179.