Add custom `expect.toHaveMembers([...])` matcher
Add a custom matcher to server tests that allows us to verify a list or set has a specific list of members/values, in any order. It is *similar* to a set comparison, but accommodates having more than of a particular value. For example, this test fails, where it would pass if it was a set comparison:

    expect(["a", "a", "b"]).toHaveMembers(["a", "b"]);

This is mainly useful for `location.availability.sources`, where we currently output the list of sources in the order they were merged, but don't want to make any *guarantees* about that.