Add incident reports in `/docs/incidents`
This adds a template and process for writing incident reports when we have outages or major infrastructural problems. It also includes actual incident reports for our first two inci
dents: 2021-06-03 (ran out of disk space for the database) and 2021-06-11 (running low on disk space for the database). It partially covers #209.