Add basic runbook/ops documentation (#248)
This adds some pretty basic ops documentation about how the current system is set up and things you might do to manage it in `docs/runbook`. We probably need to be somewhat careful about how much detail we add here since this is soon to be open-sourced. We need a more permissioned place to share actual server addresses, ARNs (not sure how sensitive we should consider those), credentials, etc.

This is the second piece of #209, along with #247.