Switch to loading task images from new ECR repos
This is follow-on 2 for #258. Now that we are pushing images to the new `univaf-*` ECR repositories, we can safely start loading images from those repositories, which is what this PR does.

This looks like a new deploy to the API server, but there are no code changes — it just has a new hash because we started pushing to the new repository in a different commit.