Rename things to “UNIVAF” (#255)
We’ve got a lot of verbosely or variously named components in this codebase, even though we colloquially refer to everything as UNIVAF. This renames all the applicable parts of the code to “UNIVAF/univaf” to organize and clarify things.

The roughest part of this may be changing the default database name. Everywhere folks are using environment variables to access their database, this should be fine, since whatever they have set locally will win out (which includes production). However, if someone is using docker-compose, they might run into issues since the environment variables are hard-coded there.

Fixes #191.