Better docs for open-source (#259)
Prepare for open-sourcing by adding a more general description to the README and removing task plans from first few weeks of the project, which are no longer relevant. This also adds other critical open-source docs, including a license, code of conduct, and contribution guide. Most of these docs are based on USDR's broader policies. Finally, the README now also has a list of contributors.

This is part of #233.