Build images less often, and only after tests
We build a lot of unnecessary docker images for commits that don't change any code involved in the build (e.g. when we update terraform configs to deploy something). This changes our CI workflow so it doesn't build the images if only the `docs`, `tarraform`, or `docker/postgis` directories changed. It also only builds after (and if) tests and linting succeeded.