Don't output to STDOUT if sending via HTTP
The loader currently streams output data on STDOUT, but when run on an automated schedule and using `--send` to send data to the API, that results in *huge* logfiles that aren't really useful for solving most problems and are instead very costly to manage. To avoid those issues, using the `--send` option now prevents data from being written to STDOUT.

Fixes #221.