2021-07-08 Dependency update for server
Snyk posted a whole mess of updates, but we should probably be updating to things that are slightly newer anyway. This should supersede those Snyk PRs. This commit does not update prettier yet, since it contains updates that will change our code formatting, and we should coordinate that across all the projects.