Improve SMART SL API, use new aggregator guidance (#299)
This vastly improves our implementation of the SMART Scheduling Links standard. It follows the new section with features designed for aggregators and also adds actual slots or daily capacity where we have them, now that we have top-level fields for that info.

- Fixes identifiers to use the new `external_ids` format.
- Simplifies location and schedule IDs.
- Adds last updated time to locations.
- Adds the `has-availability` extension on schedules instead of the wonky way we used to handle unknown availability.
- Adds `lastSourceSync` meta extension.
- Lists actual individual slots instead of one placeholder slot per location. If we don't know slots but do know daily capcity, that's used to create slots instead.
- Adds booking phone and deep link on slots.
- Adds capacity on slots.
- Includes available products and doses on each schedule.

This doesn’t create separate schedules for each product & dose combination, but we could do that in the future. It’s still technically valid, although the standard prefers that each product have a separate schedule if possible.

This also passes the two validators:
- https://infernotest.healthit.gov/community/
- https://github.com/lazau/scheduling-links-aggregator/tree/main/validator

This also adds docs about all the various particulars or caveats to our implementation.

Fixes #117.