Upgrade Sentry client in server to v6.11.0 (#328)
Note that Sentry 6.9.0 had a breaking change that needed to be accounted for here: the default value for the `autoSessionTracking` option in `Sentry.init()` changed from `false` to `true`, which causes Sentry to need explicit shutdown. Since we don’t actually need to track sessions, we’ve simply set this option explicitly to `false`.

Supersedes #318, #319, #332, #333.