Add failing test for #369
This makes some minor code changes so the loader is more friendly to testing, and fixes one minor issue where the return value of `checkAvailability` was incorrect (but all the right callbacks were getting called correctly, so it wasn't a huge problem in practice). Ultimately we should have some more expansive tests, but I just wanted to make sure the problem at hand was covered for now.