Only show in-stock products from CDC loader (#370)
This changes the CDC loader to only list products that are in stock. Fixes #369.

This adds a single test case to cover this particular situation, though it would be good to do another pass in the future to cover things more expansively. It also makes another minor fix where the return value of `checkAvailability()` was incorrect (it didn’t matter that much in production because callbacks were still correct, and that’s what usually is relied on).