HOTFIX: Bump max request body size in API server (#388)
A few locations from the PrepMod loader are too big for our current 500 kB limit, so this is the quickest and straightest solution. We might want to do a second pass on the loader where we summarize the data more briefly, since this is a lot to ship *out* and to do joins on in our own API.

Fixes #387.