Set a standard user agent header for loader requests (#239)
We now have `utils.httpClient`, which is a customized version of `got` with the correct `User-Agent` header (we can also make other future customizations here). Anywhere we make HTTP requests, we should use `utils.httpClient` instead of `got` unless there’s a very special situation.

There is also a `RELEASE` environment variable that, in production, will have the commit hash the release was built from. If release isn’t set, it falls back to using the `version` field from `package.json`. This is used to fill in the version number in the User Agent string.

Fixes #93.

Co-authored-by: Greg Sandstrom <greg@gregsandstrom.com>
Co-authored-by: Rob Brackett <rob@robbrackett.com>