Extract common SMART SL code into a module (#401)
We’ve got several SMART SL loaders we need to implement this week before VaccineSpotter completely shuts down, and this should make that a little easier by extracting the common SMART SL parsing code we currently have. I’m still not super happy with how this client works, but it’s what we’ve got. We’ll probably be doing more refactoring on this later.