Add non-zero-padded IDs for VaccineSpotter
While working on #160, I noticed a lot of differences in numbers that are zero-padded, not zero-padded, or zero-padded to different lengths between the official Kroger API and the VaccineSpotter data we have so far. To resolve the issue, this makes the VaccineSpotter loader start sending non-zero-padded IDs alongside the IDs it currently sends.