Update locations correctly when not public (#406)
It turns out we fail to find an existing location if it's non-public when an update is sent to `/update`. That seems to have lead to us creating 120 copies of the 18 non-public CVS locations we have! This should fix the issue so we actually update existing locations, even if they aren't marked as public.