Update `remove-duplicate-locations` script
This updates the remove duplicate locations script to work better with our current approach to external IDs. It also simplifies some of the more complex logic around grouping all the duplicates together, and adds new `--unpad` and `--system` options that will be useful in cleaning up some issues with CVS and Kroger (see #160, #406).

This is a follow-on to #407.