Add more details to merge plan output
When printing a merge plan, include the address and name of each location being merged. *Most* of the time, we expect two locations being merged to be pretty similar, and this is helpful in sanity checking a plan. Additionally, check the addresses for rough similarity, and output a warning if they are different. Plan output now looks like:

```
Merging into d555858a-250b-4d6c-a34b-e173e6a77f16 (7385 Wooster Pike, Cincinnati, OH 45227 // Kroger - Mariemont)
  from       e9361db1-1996-4771-a326-b3b40f3acaac (7385 WOOSTER PIKE, Mariemont, OH 45227 // Kroger Pharmacy #01400421)
  ID: kroger:01400421
  ID: kroger:1400421
  ID: univaf_v1:e9361db1-1996-4771-a326-b3b40f3acaac
  ID: vaccines_gov:1132fb34-b6fd-42cf-918e-c9d0c18743a5
  ID: vaccinespotter:9610795
  Updating fields: {"info_phone":"(513) 271-9320","info_url":"www.krogerhealth.com","meta":{"friday_hours":"9:00 AM - 8:00 PM","monday_hours":"9:00 AM - 8:00 PM","sunday_hours":"10:00 AM - 5:00 PM","tuesday_hours":"9:00 AM - 8:00 PM","saturday_hours":"10:00 AM - 6:00 PM","thursday_hours":"9:00 AM - 8:00 PM","wednesday_hours":"9:00 AM - 8:00 PM","walkins_accepted":true,"insurance_accepted":true,"time_zone":"America/New_York","vaccinespotter":{"brand":"kroger","brand_id":31,"provider":"kroger"}}}
  WARNING: Locations have different addresses.
```

The addresses and warnings were very helpful in checking the result of merges for Kroger (see https://github.com/usdigitalresponse/univaf/issues/160#issuecomment-950248690).

This expands on the existing `matchableAddress` algorithm we have in the loader utils. Unfortunately we don't have a great mechanism for sharing code between the loader and server right now, so this adds a copy of that function. :(