Re-implement Albertsons loader from VaccineSpotter (#426)
This is a port of the Albertsons source from VaccineSpotter: https://github.com/GUI/covid-vaccine-spotter/tree/main/src/providers/Albertsons. We're porting it because they're shutting down, and the Albertsons implementation is pretty simple and maintainable (it loads a JSON file that Albertsons publishes), and Albertsons is very important to our state partners.

Fixes #420.