End the updated at scourge (#431)
It turns out we missed one source when originally refactoring to `valid_at`. (The server transparently fixes this for us, but it would be nice to be able to drop that code!)

This also fixes the UI demo, which was still trying to read a non-existent `updated_at` field.

Fixes #97.