Always update a location's `external_ids`
Even if we're not updating the other properties of a location (name, address, etc.), we should always add any new external IDs we see for a location. This is a more reasonable thing than it used to be, since you can now only add to the list for a given location, and not change existing entries.