Add schemas for use in tests (#437)
This adds `expect(x).toMatchSchema(someJsonSchemaDefinition)` and `expect(x).toContainItemsMatchingSchema(someJsonSchemaDefinition)` matchers to Jest in the loader project, and should be helpful in writing better tests.

There is a very nice `jest-json-schema` package on NPM that does something similar, but it is stuck on a several-major-releases-old version of AJV (https://github.com/americanexpress/jest-json-schema/pull/26). I wanted our schemas in the loader and server to be similar, and not have some weird edge cases because of AJV versioning problems, so I wound up just implementing something custom. At least it’s not super complicated, but still makes me sad. :(

The schemas are not the same as the server because:
- They needed to be more strict. The server ones are designed to help massage acceptable data into correct data, and here, we want to be really narrow about what we do.
- The server schemas *manipulate* the data (again, aiding in cleaning up and reshaping roughly correct data), which is not what we want here.

Ideally, the server would somehow build on/extend/inherit from these more strict schemas, but that felt like a lot more work to do and other things (https://github.com/usdigitalresponse/univaf/milestone/5) are much higher priority at the moment.