Expand store support in CDC API source (#443)
This fixes Walmart/Sam's Club support in the CDC source, and also adds support for:
- Shoprite
- Stop & Shop
- Sav-On
- Genoa Healthcare
- Bartell Drug
- Meijer
- Winn Dixie
- Hannaford

This list is focused on chains that are common in our partners' states (AK, NJ, WA) and that had lots of entries in the CDC data. I also did my best to make sure these external IDs seem sensible and match up to what we’ve already got (when we already have something for these stores). One exception here is Wegmans, which I couldn’t find a reasonable way to match up and wrote a big comment about inline.

Fixes #393.