Support new pediatric data in Albertsons (#473)
Albertsons started showing pediatric data (yay!) but they did it by having separate records for adult vs. pediatric vaccines (messy for us). See the added test in this PR for a good example.

This attempts to handle that by grouping and merging locations with the same store number (or name, if lacking a store number) after formatting. (We could also group before formatting; it felt kind of like a tossup on complexity to me.)